% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs.R
\name{DIC}
\alias{DIC}
\title{Deviance Information Criterion}
\usage{
DIC(object)
}
\arguments{
\item{object}{A result from the Gibbs sampler.}
}
\value{
DIC computed from saved draws
}
\description{
A function to compute the Deviance Information Criterion (DIC) on an
\code{stcos} object.
}
\examples{
\dontrun{
out1 <- gibbs.stcos(sp, R = 10000, burn = 0, thin = 1)
DIC(out1)
}
}
\seealso{
\code{\link{gibbs.stcos}} \code{\link{gibbs.stcos.raw}}
}
